better fix

+++ segment.py
@@ -86,7 +86,7 @@ class WsUtils:
     def updateCurrentPdf(self, pageNumb, fileName):
         self.currentPdf = pageNumb
         self.fileName = fileName
-
+        
         data = copy.deepcopy(self._jsonBaseObject)
         data["contents"]["currentPdf"] = pageNumb
         data["contents"]["fileName"] = fileName
@@ -101,9 +101,9 @@ class WsUtils:
     def sendInitzDataOnConenction(self, client):
         data = copy.deepcopy(self._jsonBaseObject)
         data["contents"]["setState"] = self.getStrStateFromEnum(self.state)
-        if self.state == State.GENERATING_IMAGES:
+        if (self.state == State.GENERATING_IMAGES):
             data["contents"]["imagePages"] = self.numberOfPages
-        elif self.state == State.PROCESSING:
+        elif (self.state == State.PROCESSING):
             data["contents"]["numberOfPDFs"] = self.numberOfPDFs
             data["contents"]["currentPdf"] = self.currentPdf
             data["contents"]["fileName"] = self.fileName
@@ -150,18 +150,9 @@ def checkFile(fullfile, invalid_files):
         except:
             return False
 
-
 def move_file_to_invalid_files(sourceDirectoryPath, filename):
-    shutil.move(
-        os.path.join(sourceDirectoryPath, filename),
-        os.path.join(config["INVALID_INPUT_FOLDER"], filename),
-    )
-    print(
-        "WARNING: PDF corrupt, and moved to the invalid input folder. ("
-        + str(filename)
-        + ")"
-    )
-
+    shutil.move(os.path.join(sourceDirectoryPath, filename), os.path.join(config["INVALID_INPUT_FOLDER"], filename))
+    print("WARNING: PDF corrupt, and moved to the invalid input folder. (" + str(filename) + ")")
 
 def segment_documents(args: str):
     """
@@ -178,7 +169,7 @@ def segment_documents(args: str):
     tmp_folder = os.path.join(config["OUTPUT_FOLDER"], "tmp")
     IO_handler.folder_prep(config["OUTPUT_FOLDER"], args.clean)
     # Ghostscript converts text to images, and therfore it can not be used by pdf miner.
-    # invalid_files = gs.run_ghostscript(config["INPUT_FOLDER"])
+    #invalid_files = gs.run_ghostscript(config["INPUT_FOLDER"])
     pdf2png = Pdf2Png(3, False)
     wsUtils.setState(State.GENERATING_IMAGES)
 
@@ -210,9 +201,9 @@ def segment_documents(args: str):
         move_file_to_invalid_files(config["INPUT_FOLDER"], file)
 
     wsUtils.numberOfPDFs = len(os.listdir(config["INPUT_FOLDER"]))
-    wsUtils.updateNumberOfPdfFiles()  # Send updated page number after removed files
+    wsUtils.updateNumberOfPdfFiles() # Send updated page number after removed files
     wsUtils.setState(State.PROCESSING)
-
+    
     fileNumber = 0
     for file in os.listdir(config["INPUT_FOLDER"]):
         fileNumber += 1
@@ -221,8 +212,8 @@ def segment_documents(args: str):
             try:
                 output_path = os.path.join(
                     config["OUTPUT_FOLDER"], os.path.basename(file).replace(".pdf", "")
-                )
-
+                )             
+                 
                 print("\nGathering meta data...")
 
                 doc = fitz.open(os.path.join(config["INPUT_FOLDER"], file))
@@ -248,7 +239,6 @@ def segment_documents(args: str):
                 while True:
                     if page != last_page:
                         wsUtils.updatePageNumbers(page.value, pages)
-                        
                     if not seg_doc_process.is_alive():
                         print("Thread done!")
                         seg_doc_process.terminate()
@@ -258,16 +248,15 @@ def segment_documents(args: str):
                     if time.time() > max_time:
                         seg_doc_process.terminate()
                         print("Process: " + file + " terminated due to excessive time")
-                        # warn.warn(f"Process: {file} terminated due to excessive time", UserWarning)
-                        shutil.rmtree(output_path)
-                        break
+                        # don't try to remove multiple times
+                        if os.path.isdir(output_path):
+                            shutil.rmtree(output_path)
 
                     # Kills process if memory usage is high
                     virtual = psutil.virtual_memory()
                     if virtual.percent > 99:
                         seg_doc_process.terminate()
                         print("Memory usage above 99%. PDF file extraction killed")
-                        break
                         
                     time.sleep(0.1)  # how often to check timer
